% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_evaluation.R
\name{did_estimate}
\alias{did_estimate}
\title{Difference-in-Differences Estimation}
\usage{
did_estimate(data, outcome_var, treat_var, time_var)
}
\arguments{
\item{data}{Panel data frame}

\item{outcome_var}{Outcome variable}

\item{treat_var}{Treatment indicator}

\item{time_var}{Time period indicator}
}
\value{
DID estimate
}
\description{
Difference-in-Differences Estimation
}
