% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_analysis.R
\name{gini}
\alias{gini}
\title{Calculate Gini Coefficient}
\usage{
gini(x)
}
\arguments{
\item{x}{Numeric vector of wealth/income values}
}
\value{
Gini coefficient between 0 and 1
}
\description{
Calculate Gini Coefficient
}
\examples{
gini(c(10, 20, 30, 40, 50))
}
